--- src/harish/Test.java
@@ -1,5 +1,0 @@
-package harish;
-
-public class Test {
-
-}
