+++ src/harish/loopsdemo/Assignment12.java
@@ -1,4 +1,4 @@
-package harish;
+package harish.loopsdemo;
 
 /**
  * Assignment 12 - Print all even numbers in a user defined range
